--- a/lib/egg/context_logger.js
+++ b/lib/egg/context_logger.js
@@ -1,12 +1,11 @@
 'use strict';
 
-const { performance } = require('perf_hooks');
+const { contextFormatter } = require('../utils');
 
 /**
  * Request context Logger, itself isn't a {@link Logger}.
  */
 class ContextLogger {
-
   /**
    * @class
    * @param {Context} ctx - egg Context instance
@@ -17,48 +16,21 @@ class ContextLogger {
     this._logger = logger;
   }
 
-  get paddingMessage() {
-    const ctx = this.ctx;
-
-    // Auto record necessary request context infomation, e.g.: user id, request spend time
-    // format: '[$userId/$ip/$traceId/$use_ms $method $url]'
-    const userId = ctx.userId || '-';
-    const traceId = ctx.tracer && ctx.tracer.traceId || '-';
-    let use = 0;
-    if (ctx.performanceStarttime) {
-      use = Math.floor((performance.now() - ctx.performanceStarttime) * 1000) / 1000;
-    } else if (ctx.starttime) {
-      use = Date.now() - ctx.starttime;
-    }
-    return '[' +
-      userId + '/' +
-      ctx.ip + '/' +
-      traceId + '/' +
-      use + 'ms ' +
-      ctx.method + ' ' +
-      ctx.url +
-    ']';
-  }
-
   write(msg) {
     this._logger.write(msg);
   }
 }
 
+// logger.error()/warn()/info()/debug()
 [ 'error', 'warn', 'info', 'debug' ].forEach(level => {
   const LEVEL = level.toUpperCase();
-  ContextLogger.prototype[level] = function() {
+  ContextLogger.prototype[level] = function(...args) {
     const meta = {
       formatter: contextFormatter,
-      paddingMessage: this.paddingMessage,
       ctx: this.ctx,
     };
-    this._logger.log(LEVEL, arguments, meta);
+    this._logger.log(LEVEL, args, meta);
   };
 });
 
 module.exports = ContextLogger;
-
-function contextFormatter(meta) {
-  return meta.date + ' ' + meta.level + ' ' + meta.pid + ' ' + meta.paddingMessage + ' ' + meta.message;
-}