--- a/lib/egg/context_logger.js
+++ b/lib/egg/context_logger.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { contextFormatter } = require('../utils');
+const { contextFormatter, defaultContextPaddingMessage } = require('../utils');
 
 /**
  * Request context Logger, itself isn't a {@link Logger}.
@@ -16,6 +16,11 @@ class ContextLogger {
     this._logger = logger;
   }
 
+  // sub class can override this getter
+  get paddingMessage() {
+    return defaultContextPaddingMessage(this.ctx);
+  }
+
   write(msg) {
     this._logger.write(msg);
   }
@@ -27,6 +32,7 @@ class ContextLogger {
   ContextLogger.prototype[level] = function(...args) {
     const meta = {
       formatter: contextFormatter,
+      paddingMessage: this.paddingMessage,
       ctx: this.ctx,
     };
     this._logger.log(LEVEL, args, meta);