--- a/lib/egg/context_logger.js
+++ b/lib/egg/context_logger.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { contextFormatter, defaultContextPaddingMessage } = require('../utils');
+const { defaultContextPaddingMessage } = require('../utils');
 
 /**
  * Request context Logger, itself isn't a {@link Logger}.
@@ -31,7 +31,6 @@ class ContextLogger {
   const LEVEL = level.toUpperCase();
   ContextLogger.prototype[level] = function(...args) {
     const meta = {
-      formatter: contextFormatter,
       paddingMessage: this.paddingMessage,
       ctx: this.ctx,
     };