--- a/lib/egg/context_logger.js
+++ b/lib/egg/context_logger.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const depd = require('depd')('egg-logger');
 const { defaultContextPaddingMessage } = require('../utils');
 
 /**
@@ -14,6 +15,7 @@ class ContextLogger {
   constructor(ctx, logger) {
     this.ctx = ctx;
     this._logger = logger;
+    depd('EggContextLogger is deprecated, use the EggLogger directly');
   }
 
   // sub class can override this getter