--- a/lib/egg/logger.js
+++ b/lib/egg/logger.js
@@ -26,6 +26,7 @@ class EggLogger extends Logger {
    *  - {Boolean} [buffer] - use {@link FileBufferTransport} or not
    *  - {String} [eol] - end of line char
    *  - {String} [concentrateError] - whether write error logger to common-error.log, `duplicate` / `redirect` / `ignore`
+   *  - {AsyncLocalStorage} [localStorage] - AsyncLocalStorage instance to get current ctx
    */
   constructor(options) {
     super(options);
@@ -47,6 +48,7 @@ class EggLogger extends Logger {
         contextFormatter: this.options.contextFormatter,
         flushInterval: this.options.flushInterval,
         eol: this.options.eol,
+        localStorage: this.options.localStorage,
       });
       this.set('file', fileTransport);
     }
@@ -59,6 +61,7 @@ class EggLogger extends Logger {
         flushInterval: this.options.flushInterval,
         json: true,
         eol: this.options.eol,
+        localStorage: this.options.localStorage,
       });
       this.set('jsonFile', jsonFileTransport);
     }
@@ -68,6 +71,7 @@ class EggLogger extends Logger {
       formatter: utils.consoleFormatter,
       contextFormatter: this.options.contextFormatter,
       eol: this.options.eol,
+      localStorage: this.options.localStorage,
     });
     this.set('console', consoleTransport);
   }