--- a/lib/level.js
+++ b/lib/level.js
@@ -5,28 +5,28 @@ exports.ALL = -Infinity;
 /**
  * Debug log for execute tracing
  *
- * @method Logger#debug
+ * @function Logger#debug
  * @param {String} msg - log message
  */
 exports.DEBUG = 0;
 
 /**
  * Normal infomaction loging
- * @method Logger#info
+ * @function Logger#info
  * @param {String} msg - log message
  */
 exports.INFO = 1;
 
 /**
  * Warning information loging
- * @method Logger#warn
+ * @function Logger#warn
  * @param {String} msg - log message
  */
 exports.WARN = 2;
 
 /**
  * Error or exception loging
- * @method Logger#error
+ * @function Logger#error
  * @param {String|Error} msg - log message or error instance
  */
 exports.ERROR = 3;