--- a/lib/logger.js
+++ b/lib/logger.js
@@ -4,7 +4,6 @@ const util = require('util');
 const utils = require('./utils');
 const depd = require('depd')('egg-logger');
 
-
 /**
  * Base class for all sub Logger class.
  * It extends Map, and can contains multi {@link Transport}