--- a/lib/logger.js
+++ b/lib/logger.js
@@ -164,8 +164,8 @@ class Logger extends Map {
 
 [ 'error', 'warn', 'info', 'debug' ].forEach(level => {
   const LEVEL = level.toUpperCase();
-  Logger.prototype[level] = function() {
-    this.log(LEVEL, arguments);
+  Logger.prototype[level] = function(...args) {
+    this.log(LEVEL, args);
   };
 });