--- a/lib/transports/file.js
+++ b/lib/transports/file.js
@@ -3,7 +3,7 @@
 const fs = require('fs');
 const path = require('path');
 const assert = require('assert');
-const mkdirp = require('mkdirp');
+const { mkdirSync } = require('fs');
 const utility = require('utility');
 const depd = require('depd')('egg-logger');
 const Transport = require('./transport');
@@ -102,7 +102,7 @@ class FileTransport extends Transport {
    * @private
    */
   _createStream() {
-    mkdirp.sync(path.dirname(this.options.file));
+    mkdirSync(path.dirname(this.options.file), { recursive: true });
     const stream = fs.createWriteStream(this.options.file, { flags: 'a' });
 
     const onError = err => {