--- a/lib/transports/transport.js
+++ b/lib/transports/transport.js
@@ -21,6 +21,7 @@ class Transport {
    * - {Boolean} [json = false] - log format is json or not
    * - {String} [encoding = utf8] - log encodeing, see {@link https://github.com/ashtuchkin/iconv-lite#supported-encodings}
    * - {String} [eol = os.EOL] - end of line
+   * - {AsyncLocalStorage} [localStorage] - AsyncLocalStorage instance to get current ctx
    */
   constructor(options) {
     this.options = utils.assign(this.defaults, options);
@@ -39,6 +40,7 @@ class Transport {
       json: false,
       encoding: 'utf8',
       eol: os.EOL,
+      // localStorage: null,
     };
   }
 
@@ -100,6 +102,12 @@ class Transport {
    *   - other encoding return Buffer
    */
   log(level, args, meta) {
+    if (!meta?.ctx && this.options.localStorage) {
+      const ctx = this.options.localStorage.getStore();
+      if (ctx) {
+        meta = { ...meta, ctx };
+      }
+    }
     return utils.format(level, args, meta, this.options);
   }