--- a/lib/utils.js
+++ b/lib/utils.js
@@ -2,12 +2,13 @@
 
 const os = require('os');
 const util = require('util');
+const { performance } = require('perf_hooks');
 const chalk = require('chalk');
 const utility = require('utility');
 const iconv = require('iconv-lite');
-const levels = require('./level');
 const circularJSON = require('circular-json-for-egg');
 const { FrameworkBaseError, FrameworkErrorFormater } = require('egg-errors');
+const levels = require('./level');
 
 const hostname = os.hostname();
 const duartionRegexp = /([0-9]+ms)/g;
@@ -33,6 +34,35 @@ module.exports = {
   // default format
   defaultFormatter(meta) {
     return meta.date + ' ' + meta.level + ' ' + meta.pid + ' ' + meta.message;
+  },
+
+  // format with ctx and ctx.tracer
+  // Auto record necessary request context infomation, e.g.: user id, request spend time
+  // format: '[$userId/$ip/$traceId/$use_ms $method $url]'
+  contextFormatter(meta) {
+    let paddingMessage = ' ';
+    const ctx = meta.ctx;
+    if (ctx) {
+      // Auto record necessary request context infomation, e.g.: user id, request spend time
+      // format: '[$userId/$ip/$traceId/$use_ms $method $url]'
+      const userId = ctx.userId || '-';
+      const traceId = ctx.tracer?.traceId || '-';
+      let use = 0;
+      if (ctx.performanceStarttime) {
+        use = Math.floor((performance.now() - ctx.performanceStarttime) * 1000) / 1000;
+      } else if (ctx.starttime) {
+        use = Date.now() - ctx.starttime;
+      }
+      paddingMessage = ' [' +
+        userId + '/' +
+        ctx.ip + '/' +
+        traceId + '/' +
+        use + 'ms ' +
+        ctx.method + ' ' +
+        ctx.url +
+      '] ';
+    }
+    return meta.date + ' ' + meta.level + ' ' + meta.pid + paddingMessage + meta.message;
   },
 
   // output to Terminal format