--- a/lib/utils.js
+++ b/lib/utils.js
@@ -36,31 +36,41 @@ module.exports = {
     return meta.date + ' ' + meta.level + ' ' + meta.pid + ' ' + meta.message;
   },
 
+  defaultContextPaddingMessage(ctx) {
+    // Auto record necessary request context infomation, e.g.: user id, request spend time
+    // format: '[$userId/$ip/$traceId/$use_ms $method $url]'
+    const userId = ctx.userId || '-';
+    const traceId = ctx.tracer?.traceId || '-';
+    let use = 0;
+    if (ctx.performanceStarttime) {
+      use = Math.floor((performance.now() - ctx.performanceStarttime) * 1000) / 1000;
+    } else if (ctx.starttime) {
+      use = Date.now() - ctx.starttime;
+    }
+    return '[' +
+      userId + '/' +
+      ctx.ip + '/' +
+      traceId + '/' +
+      use + 'ms ' +
+      ctx.method + ' ' +
+      ctx.url +
+    ']';
+  },
+
   // format with ctx and ctx.tracer
   // Auto record necessary request context infomation, e.g.: user id, request spend time
   // format: '[$userId/$ip/$traceId/$use_ms $method $url]'
   contextFormatter(meta) {
     let paddingMessage = ' ';
-    const ctx = meta.ctx;
-    if (ctx) {
-      // Auto record necessary request context infomation, e.g.: user id, request spend time
-      // format: '[$userId/$ip/$traceId/$use_ms $method $url]'
-      const userId = ctx.userId || '-';
-      const traceId = ctx.tracer?.traceId || '-';
-      let use = 0;
-      if (ctx.performanceStarttime) {
-        use = Math.floor((performance.now() - ctx.performanceStarttime) * 1000) / 1000;
-      } else if (ctx.starttime) {
-        use = Date.now() - ctx.starttime;
+    // try to use the exists paddingMessage first
+    if (meta.paddingMessage) {
+      paddingMessage = ` ${meta.paddingMessage} `;
+    } else {
+      // gen the default ctx paddingMessage
+      const ctx = meta.ctx;
+      if (ctx) {
+        paddingMessage = ` ${module.exports.defaultContextPaddingMessage(ctx)} `;
       }
-      paddingMessage = ' [' +
-        userId + '/' +
-        ctx.ip + '/' +
-        traceId + '/' +
-        use + 'ms ' +
-        ctx.method + ' ' +
-        ctx.url +
-      '] ';
     }
     return meta.date + ' ' + meta.level + ' ' + meta.pid + paddingMessage + meta.message;
   },