--- a/lib/utils.js
+++ b/lib/utils.js
@@ -31,11 +31,6 @@ module.exports = {
     }
   },
 
-  // default format
-  defaultFormatter(meta) {
-    return meta.date + ' ' + meta.level + ' ' + meta.pid + ' ' + meta.message;
-  },
-
   defaultContextPaddingMessage(ctx) {
     // Auto record necessary request context infomation, e.g.: user id, request spend time
     // format: '[$userId/$ip/$traceId/$use_ms $method $url]'
@@ -60,7 +55,7 @@ module.exports = {
   // format with ctx and ctx.tracer
   // Auto record necessary request context infomation, e.g.: user id, request spend time
   // format: '[$userId/$ip/$traceId/$use_ms $method $url]'
-  contextFormatter(meta) {
+  defaultFormatter(meta) {
     let paddingMessage = ' ';
     // try to use the exists paddingMessage first
     if (meta.paddingMessage) {