--- a/test/file_buffer_heap.js
+++ b/test/file_buffer_heap.js
@@ -3,7 +3,7 @@
 const fs = require('fs');
 // need to install heapdump by yourself
 const heapdump = require('heapdump');
-const sleep = require('mz-modules/sleep');
+const { sleep } = require('./utils');
 const Logger = require('../index').Logger;
 const FileBufferTransport = require('../index').FileBufferTransport;