--- a/test/lib/egg/console_logger.test.js
+++ b/test/lib/egg/console_logger.test.js
@@ -3,14 +3,14 @@
 const path = require('path');
 const coffee = require('coffee');
 const mm = require('mm');
-const rimraf = require('rimraf');
+const { rimraf } = require('../../utils');
 
 describe('test/egg/console_logger.test.js', () => {
   const consoleLoggerFile = path.join(__dirname, '../../fixtures/egg_console_logger.js');
   const tmp = path.join(__dirname, '../../fixtures/tmp');
 
-  afterEach(() => {
-    rimraf.sync(tmp);
+  afterEach(async () => {
+    await rimraf(tmp);
   });
   afterEach(mm.restore);