--- a/test/lib/egg/custom_logger.test.js
+++ b/test/lib/egg/custom_logger.test.js
@@ -1,10 +1,10 @@
 'use strict';
 
 const assert = require('assert');
-const { fs } = require('mz');
+const { fs } = require('fs/promises');
 const path = require('path');
 const coffee = require('coffee');
-const { rimraf } = require('mz-modules');
+const { rimraf } = require('../../utils');
 
 describe('test/egg/custom_logger.test.js', () => {
   const loggerFile = path.join(__dirname, '../../fixtures/egg_custom_logger.js');
@@ -14,25 +14,25 @@ describe('test/egg/custom_logger.test.js', () => {
   beforeEach(() => rimraf(path.dirname(filePath)));
   afterEach(() => rimraf(path.dirname(filePath)));
 
-  it('should format work', function*() {
+  it('should format work', async () => {
     const options = {
       file: filePath,
       level: 'WARN',
     };
-    yield coffee.fork(loggerFile, [ JSON.stringify(options) ]).end();
-    const log = yield fs.readFile(filePath, 'utf-8');
+    await coffee.fork(loggerFile, [ JSON.stringify(options) ]).end();
+    const log = await fs.readFile(filePath, 'utf-8');
     assert(log.match(/\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2},\d{3} ERROR \d+ error foo\n/));
   });
 
-  it('should support relative path', function* () {
+  it('should support relative path', async () => {
     const options = {
       dir: tmpDir,
       file: 'relative.log',
       level: 'WARN',
     };
-    yield coffee.fork(loggerFile, [ JSON.stringify(options) ]).end();
+    await coffee.fork(loggerFile, [ JSON.stringify(options) ]).end();
 
-    const log = yield fs.readFile(path.join(tmpDir, options.file), 'utf-8');
+    const log = await fs.readFile(path.join(tmpDir, options.file), 'utf-8');
     assert(log.match(/\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2},\d{3} ERROR \d+ error foo\n/));
   });
 });