- Drop Node.js <= 14 support
- deprecated EggContextLogger
- Upgrade chalk@4