@fengmk2 2022-12-10
## Benchmark

### Before

```
  node version: v18.12.1, date: Sat Dec 10 2022 20:16:26 GMT+0800 (中国标准时间)
  Starting...
  10 tests completed.

  logger.error(err)                                           x   419,453 ops/sec ±0.67% (93 runs sampled)
  logger.info(message)                                        x 2,078,658 ops/sec ±0.88% (95 runs sampled)
  logger.write(message)                                       x 8,510,356 ops/sec ±0.68% (96 runs sampled)
  contextLogger.info(message)                                 x 1,397,042 ops/sec ±1.08% (96 runs sampled)
  performanceStarttime: contextLogger.info(message)           x 1,136,481 ops/sec ±1.40% (92 runs sampled)
  logger.error(err) with JSON                                 x   133,504 ops/sec ±1.09% (95 runs sampled)
  logger.info(message) with JSON                              x   532,361 ops/sec ±1.53% (98 runs sampled)
  logger.write(message) with JSON                             x 4,166,263 ops/sec ±2.06% (90 runs sampled)
  contextLogger.info(message) with JSON                       x   436,971 ops/sec ±3.02% (84 runs sampled)
  performanceStarttime: contextLogger.info(message) with JSON x   434,387 ops/sec ±1.42% (92 runs sampled)
```

### After

```
  node version: v18.12.1, date: Sat Dec 10 2022 20:18:07 GMT+0800 (中国标准时间)
  Starting...
  10 tests completed.

  logger.error(err)                                           x   417,021 ops/sec ±0.62% (97 runs sampled)
  logger.info(message)                                        x 2,091,101 ops/sec ±0.57% (98 runs sampled)
  logger.write(message)                                       x 8,428,986 ops/sec ±0.76% (97 runs sampled)
  contextLogger.info(message)                                 x 1,382,615 ops/sec ±1.01% (90 runs sampled)
  performanceStarttime: contextLogger.info(message)           x 1,143,361 ops/sec ±0.95% (96 runs sampled)
  logger.error(err) with JSON                                 x   131,539 ops/sec ±1.11% (96 runs sampled)
  logger.info(message) with JSON                              x   530,138 ops/sec ±1.31% (93 runs sampled)
  logger.write(message) with JSON                             x 4,326,800 ops/sec ±1.12% (89 runs sampled)
  contextLogger.info(message) with JSON                       x   489,743 ops/sec ±1.36% (93 runs sampled)
  performanceStarttime: contextLogger.info(message) with JSON x   448,977 ops/sec ±1.10% (95 runs sampled)
```